+++ cmd/frontend/graphqlbackend/graphqlbackend.go
@@ -2,6 +2,7 @@ package graphqlbackend
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"log"
 	"strconv"
@@ -18,6 +19,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/db"
 	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/errcode"
 	sgtrace "github.com/sourcegraph/sourcegraph/internal/trace"
 	"github.com/sourcegraph/sourcegraph/internal/trace/ot"
@@ -29,7 +31,7 @@ var graphqlFieldHistogram = prometheus.NewHistogramVec(prometheus.HistogramOpts{
 	Name:      "field_seconds",
 	Help:      "GraphQL field resolver latencies in seconds.",
 	Buckets:   []float64{0.01, 0.02, 0.05, 0.1, 0.2, 0.5, 1, 2, 5, 10, 30},
-}, []string{"type", "field", "error", "source"})
+}, []string{"type", "field", "error", "source", "request_name"})
 
 var codeIntelSearchHistogram = prometheus.NewHistogramVec(prometheus.HistogramOpts{
 	Namespace: "src",
@@ -49,6 +51,7 @@ type prometheusTracer struct {
 }
 
 func (prometheusTracer) TraceQuery(ctx context.Context, queryString string, operationName string, variables map[string]interface{}, varTypes map[string]*introspection.Type) (context.Context, trace.TraceQueryFinishFunc) {
+	start := time.Now()
 	var finish trace.TraceQueryFinishFunc
 	if ot.ShouldTrace(ctx) {
 		ctx, finish = trace.OpenTracingTracer{}.TraceQuery(ctx, queryString, operationName, variables, varTypes)
@@ -90,6 +93,23 @@ VARIABLES
 		if finish != nil {
 			finish(err)
 		}
+		d := time.Since(start)
+		if v := conf.Get().ObservabilityLogSlowGraphQLRequests; v != 0 && d.Milliseconds() > int64(v) {
+			encodedVariables, _ := json.Marshal(variables)
+			log15.Warn("slow GraphQL request", "time", d, "name", requestName, "user", currentUserName, "source", requestSource, "error", err, "variables", string(encodedVariables))
+			if requestName == "unknown" {
+				log.Printf(`logging complete query for slow GraphQL request above time=%v name=%s user=%s source=%s error=%v:
+QUERY
+-----
+%s
+
+VARIABLES
+---------
+%s
+
+`, d, requestName, currentUserName, requestSource, err, queryString, encodedVariables)
+			}
+		}
 	}
 }
 
@@ -102,7 +122,13 @@ func (prometheusTracer) TraceField(ctx context.Context, label, typeName, fieldNa
 	start := time.Now()
 	return ctx, func(err *gqlerrors.QueryError) {
 		isErrStr := strconv.FormatBool(err != nil)
-		graphqlFieldHistogram.WithLabelValues(typeName, prometheusFieldName(typeName, fieldName), isErrStr, string(sgtrace.RequestSource(ctx))).Observe(time.Since(start).Seconds())
+		graphqlFieldHistogram.WithLabelValues(
+			prometheusTypeName(typeName),
+			prometheusFieldName(typeName, fieldName),
+			isErrStr,
+			string(sgtrace.RequestSource(ctx)),
+			prometheusGraphQLRequestName(sgtrace.GraphQLRequestName(ctx)),
+		).Observe(time.Since(start).Seconds())
 
 		origin := sgtrace.RequestOrigin(ctx)
 		if origin != "unknown" && (fieldName == "search" || fieldName == "lsif") {
@@ -115,146 +141,141 @@ func (prometheusTracer) TraceField(ctx context.Context, label, typeName, fieldNa
 	}
 }
 
-var whitelistedPrometheusFieldNames = [][2]string{
-	{"AccessTokenConnection", "nodes"},
-	{"File", "isDirectory"},
-	{"File", "name"},
-	{"File", "path"},
-	{"File", "repository"},
-	{"File", "url"},
-	{"File2", "content"},
-	{"File2", "externalURLs"},
-	{"File2", "highlight"},
-	{"File2", "isDirectory"},
-	{"File2", "richHTML"},
-	{"File2", "url"},
-	{"FileDiff", "hunks"},
-	{"FileDiff", "internalID"},
-	{"FileDiff", "mostRelevantFile"},
-	{"FileDiff", "newPath"},
-	{"FileDiff", "oldPath"},
-	{"FileDiff", "stat"},
-	{"FileDiffConnection", "diffStat"},
-	{"FileDiffConnection", "nodes"},
-	{"FileDiffConnection", "pageInfo"},
-	{"FileDiffConnection", "totalCount"},
-	{"FileDiffHunk", "body"},
-	{"FileDiffHunk", "newRange"},
-	{"FileDiffHunk", "oldNoNewlineAt"},
-	{"FileDiffHunk", "oldRange"},
-	{"FileDiffHunk", "section"},
-	{"FileDiffHunkRange", "lines"},
-	{"FileDiffHunkRange", "Line"},
-	{"FileMatch", "file"},
-	{"FileMatch", "limitHit"},
-	{"FileMatch", "lineMatches"},
-	{"FileMatch", "repository"},
-	{"FileMatch", "revSpec"},
-	{"FileMatch", "symbols"},
-	{"GitBlob", "blame"},
-	{"GitBlob", "commit"},
-	{"GitBlob", "content"},
-	{"GitBlob", "lsif"},
-	{"GitBlob", "path"},
-	{"GitBlob", "repository"},
-	{"GitBlob", "url"},
-	{"GitCommit", "abbreviatedOID"},
-	{"GitCommit", "ancestors"},
-	{"GitCommit", "author"},
-	{"GitCommit", "blob"},
-	{"GitCommit", "body"},
-	{"GitCommit", "canonicalURL"},
-	{"GitCommit", "committer"},
-	{"GitCommit", "externalURLs"},
-	{"GitCommit", "file"},
-	{"GitCommit", "id"},
-	{"GitCommit", "message"},
-	{"GitCommit", "oid"},
-	{"GitCommit", "parents"},
-	{"GitCommit", "repository"},
-	{"GitCommit", "subject"},
-	{"GitCommit", "symbols"},
-	{"GitCommit", "tree"},
-	{"GitCommit", "url"},
-	{"GitCommitConnection", "nodes"},
-	{"GitRefConnection", "nodes"},
-	{"GitTree", "canonicalURL"},
-	{"GitTree", "entries"},
-	{"GitTree", "files"},
-	{"GitTree", "isRoot"},
-	{"GitTree", "url"},
-	{"Mutation", "configurationMutation"},
-	{"Mutation", "createOrganization"},
-	{"Mutation", "logEvent"},
-	{"Mutation", "logUserEvent"},
-	{"Query", "clientConfiguration"},
-	{"Query", "currentUser"},
-	{"Query", "discussionThreads"},
-	{"Query", "dotcom"},
-	{"Query", "extensionRegistry"},
-	{"Query", "highlightCode"},
-	{"Query", "node"},
-	{"Query", "organization"},
-	{"Query", "repositories"},
-	{"Query", "repository"},
-	{"Query", "repositoryRedirect"},
-	{"Query", "search"},
-	{"Query", "settingsSubject"},
-	{"Query", "site"},
-	{"Query", "user"},
-	{"Query", "viewerConfiguration"},
-	{"Query", "viewerSettings"},
-	{"RegistryExtensionConnection", "nodes"},
-	{"Repository", "cloneInProgress"},
-	{"Repository", "commit"},
-	{"Repository", "comparison"},
-	{"Repository", "gitRefs"},
-	{"RepositoryComparison", "commits"},
-	{"RepositoryComparison", "fileDiffs"},
-	{"RepositoryComparison", "range"},
-	{"RepositoryConnection", "nodes"},
-	{"Search", "results"},
-	{"Search", "suggestions"},
-	{"SearchAlert", "description"},
-	{"SearchAlert", "proposedQueries"},
-	{"SearchAlert", "title"},
-	{"SearchFilter", "count"},
-	{"SearchFilter", "kind"},
-	{"SearchFilter", "label"},
-	{"SearchFilter", "limitHit"},
-	{"SearchFilter", "value"},
-	{"SearchQueryDescription", "description"},
-	{"SearchQueryDescription", "query"},
-	{"SearchResultMatch", "body"},
-	{"SearchResultMatch", "highlights"},
-	{"SearchResultMatch", "url"},
-	{"SearchResults", "alert"},
-	{"SearchResults", "approximateResultCount"},
-	{"SearchResults", "cloning"},
-	{"SearchResults", "dynamicFilters"},
-	{"SearchResults", "elapsedMilliseconds"},
-	{"SearchResults", "indexUnavailable"},
-	{"SearchResults", "limitHit"},
-	{"SearchResults", "matchCount"},
-	{"SearchResults", "missing"},
-	{"SearchResults", "repositoriesCount"},
-	{"SearchResults", "results"},
-	{"SearchResults", "timedout"},
-	{"SettingsCascade", "final"},
-	{"SettingsMutation", "editConfiguration"},
-	{"SettingsSubject", "latestSettings"},
-	{"SettingsSubject", "settingsCascade"},
-	{"Signature", "date"},
-	{"Signature", "person"},
-	{"Site", "alerts"},
-	{"SymbolConnection", "nodes"},
-	{"TreeEntry", "isDirectory"},
-	{"TreeEntry", "isSingleChild"},
-	{"TreeEntry", "name"},
-	{"TreeEntry", "path"},
-	{"TreeEntry", "submodule"},
-	{"TreeEntry", "url"},
-	{"UserConnection", "nodes"},
+var whitelistedPrometheusFieldNames = map[[2]string]struct{}{
+	{"AccessTokenConnection", "nodes"}:          {},
+	{"File", "isDirectory"}:                     {},
+	{"File", "name"}:                            {},
+	{"File", "path"}:                            {},
+	{"File", "repository"}:                      {},
+	{"File", "url"}:                             {},
+	{"File2", "content"}:                        {},
+	{"File2", "externalURLs"}:                   {},
+	{"File2", "highlight"}:                      {},
+	{"File2", "isDirectory"}:                    {},
+	{"File2", "richHTML"}:                       {},
+	{"File2", "url"}:                            {},
+	{"FileDiff", "hunks"}:                       {},
+	{"FileDiff", "internalID"}:                  {},
+	{"FileDiff", "mostRelevantFile"}:            {},
+	{"FileDiff", "newPath"}:                     {},
+	{"FileDiff", "oldPath"}:                     {},
+	{"FileDiff", "stat"}:                        {},
+	{"FileDiffConnection", "diffStat"}:          {},
+	{"FileDiffConnection", "nodes"}:             {},
+	{"FileDiffConnection", "pageInfo"}:          {},
+	{"FileDiffConnection", "totalCount"}:        {},
+	{"FileDiffHunk", "body"}:                    {},
+	{"FileDiffHunk", "newRange"}:                {},
+	{"FileDiffHunk", "oldNoNewlineAt"}:          {},
+	{"FileDiffHunk", "oldRange"}:                {},
+	{"FileDiffHunk", "section"}:                 {},
+	{"FileDiffHunkRange", "lines"}:              {},
+	{"FileDiffHunkRange", "Line"}:               {},
+	{"FileMatch", "file"}:                       {},
+	{"FileMatch", "limitHit"}:                   {},
+	{"FileMatch", "lineMatches"}:                {},
+	{"FileMatch", "repository"}:                 {},
+	{"FileMatch", "revSpec"}:                    {},
+	{"FileMatch", "symbols"}:                    {},
+	{"GitBlob", "blame"}:                        {},
+	{"GitBlob", "commit"}:                       {},
+	{"GitBlob", "content"}:                      {},
+	{"GitBlob", "lsif"}:                         {},
+	{"GitBlob", "path"}:                         {},
+	{"GitBlob", "repository"}:                   {},
+	{"GitBlob", "url"}:                          {},
+	{"GitCommit", "abbreviatedOID"}:             {},
+	{"GitCommit", "ancestors"}:                  {},
+	{"GitCommit", "author"}:                     {},
+	{"GitCommit", "blob"}:                       {},
+	{"GitCommit", "body"}:                       {},
+	{"GitCommit", "canonicalURL"}:               {},
+	{"GitCommit", "committer"}:                  {},
+	{"GitCommit", "externalURLs"}:               {},
+	{"GitCommit", "file"}:                       {},
+	{"GitCommit", "id"}:                         {},
+	{"GitCommit", "message"}:                    {},
+	{"GitCommit", "oid"}:                        {},
+	{"GitCommit", "parents"}:                    {},
+	{"GitCommit", "repository"}:                 {},
+	{"GitCommit", "subject"}:                    {},
+	{"GitCommit", "symbols"}:                    {},
+	{"GitCommit", "tree"}:                       {},
+	{"GitCommit", "url"}:                        {},
+	{"GitCommitConnection", "nodes"}:            {},
+	{"GitRefConnection", "nodes"}:               {},
+	{"GitTree", "canonicalURL"}:                 {},
+	{"GitTree", "entries"}:                      {},
+	{"GitTree", "files"}:                        {},
+	{"GitTree", "isRoot"}:                       {},
+	{"GitTree", "url"}:                          {},
+	{"Mutation", "configurationMutation"}:       {},
+	{"Mutation", "createOrganization"}:          {},
+	{"Mutation", "logEvent"}:                    {},
+	{"Mutation", "logUserEvent"}:                {},
+	{"Query", "clientConfiguration"}:            {},
+	{"Query", "currentUser"}:                    {},
+	{"Query", "discussionThreads"}:              {},
+	{"Query", "dotcom"}:                         {},
+	{"Query", "extensionRegistry"}:              {},
+	{"Query", "highlightCode"}:                  {},
+	{"Query", "node"}:                           {},
+	{"Query", "organization"}:                   {},
+	{"Query", "repositories"}:                   {},
+	{"Query", "repository"}:                     {},
+	{"Query", "repositoryRedirect"}:             {},
+	{"Query", "search"}:                         {},
+	{"Query", "settingsSubject"}:                {},
+	{"Query", "site"}:                           {},
+	{"Query", "user"}:                           {},
+	{"Query", "viewerConfiguration"}:            {},
+	{"Query", "viewerSettings"}:                 {},
+	{"RegistryExtensionConnection", "nodes"}:    {},
+	{"Repository", "cloneInProgress"}:           {},
+	{"Repository", "commit"}:                    {},
+	{"Repository", "comparison"}:                {},
+	{"Repository", "gitRefs"}:                   {},
+	{"RepositoryComparison", "commits"}:         {},
+	{"RepositoryComparison", "fileDiffs"}:       {},
+	{"RepositoryComparison", "range"}:           {},
+	{"RepositoryConnection", "nodes"}:           {},
+	{"Search", "results"}:                       {},
+	{"Search", "suggestions"}:                   {},
+	{"SearchAlert", "description"}:              {},
+	{"SearchAlert", "proposedQueries"}:          {},
+	{"SearchAlert", "title"}:                    {},
+	{"SearchQueryDescription", "description"}:   {},
+	{"SearchQueryDescription", "query"}:         {},
+	{"SearchResultMatch", "body"}:               {},
+	{"SearchResultMatch", "highlights"}:         {},
+	{"SearchResultMatch", "url"}:                {},
+	{"SearchResults", "alert"}:                  {},
+	{"SearchResults", "approximateResultCount"}: {},
+	{"SearchResults", "cloning"}:                {},
+	{"SearchResults", "dynamicFilters"}:         {},
+	{"SearchResults", "elapsedMilliseconds"}:    {},
+	{"SearchResults", "indexUnavailable"}:       {},
+	{"SearchResults", "limitHit"}:               {},
+	{"SearchResults", "matchCount"}:             {},
+	{"SearchResults", "missing"}:                {},
+	{"SearchResults", "repositoriesCount"}:      {},
+	{"SearchResults", "results"}:                {},
+	{"SearchResults", "timedout"}:               {},
+	{"SettingsCascade", "final"}:                {},
+	{"SettingsMutation", "editConfiguration"}:   {},
+	{"SettingsSubject", "latestSettings"}:       {},
+	{"SettingsSubject", "settingsCascade"}:      {},
+	{"Signature", "date"}:                       {},
+	{"Signature", "person"}:                     {},
+	{"Site", "alerts"}:                          {},
+	{"SymbolConnection", "nodes"}:               {},
+	{"TreeEntry", "isDirectory"}:                {},
+	{"TreeEntry", "isSingleChild"}:              {},
+	{"TreeEntry", "name"}:                       {},
+	{"TreeEntry", "path"}:                       {},
+	{"TreeEntry", "submodule"}:                  {},
+	{"TreeEntry", "url"}:                        {},
+	{"UserConnection", "nodes"}:                 {},
 }
 
 // prometheusFieldName reduces the cardinality of GraphQL field names to make it suitable
@@ -262,11 +283,50 @@ var whitelistedPrometheusFieldNames = [][2]string{
 //
 // See https://github.com/sourcegraph/sourcegraph/issues/9895
 func prometheusFieldName(typeName, fieldName string) string {
-	for _, pair := range whitelistedPrometheusFieldNames {
-		t, f := pair[0], pair[1]
-		if t == typeName && f == fieldName {
-			return fieldName
-		}
+	if _, ok := whitelistedPrometheusFieldNames[[2]string{typeName, fieldName}]; ok {
+		return fieldName
+	}
+	return "other"
+}
+
+var blacklistedPrometheusTypeNames = map[string]struct{}{
+	"__Type":                                 {},
+	"__Schema":                               {},
+	"__InputValue":                           {},
+	"__Field":                                {},
+	"__EnumValue":                            {},
+	"__Directive":                            {},
+	"UserEmail":                              {},
+	"UpdateSettingsPayload":                  {},
+	"ExtensionRegistryCreateExtensionResult": {},
+	"Range":                                  {},
+	"LineMatch":                              {},
+	"DiffStat":                               {},
+	"DiffHunk":                               {},
+	"DiffHunkRange":                          {},
+	"FileDiffResolver":                       {},
+}
+
+// prometheusTypeName reduces the cardinality of GraphQL type names to make it
+// suitable for use in a Prometheus metric. This is a blacklist of type names
+// which involve non-complex calculations in the GraphQL backend and thus are
+// not worth tracking. You can find a complete list of the ones Prometheus is
+// currently tracking via:
+//
+// 	sum by (type)(src_graphql_field_seconds_count)
+//
+func prometheusTypeName(typeName string) string {
+	if _, ok := blacklistedPrometheusTypeNames[typeName]; ok {
+		return "other"
+	}
+	return typeName
+}
+
+// prometheusGraphQLRequestName is a whitelist of GraphQL request names (e.g. /.api/graphql?Foobar)
+// to include in a Prometheus metric. Be extremely careful
+func prometheusGraphQLRequestName(requestName string) string {
+	if requestName == "CodeIntelSearch" {
+		return requestName
 	}
 	return "other"
 }