+++ dev/check/all.sh
@@ -25,6 +25,7 @@ CHECKS=(
   ./check-owners.sh
   ./shfmt.sh
   ./shellcheck.sh
+  ./licenses.sh
 )
 
 echo "--- 🚨 Buildkite's timing information is misleading! Only consider the job timing that's printed after 'done'"